[feat] Add 'Statistics' to CoreData & Etc * Coredata   * 사용자통계(Statistics) 테이블 추가 * Object   * 'StatisticsObject' 추가 * Etc   * Int64 or Int16 변수, 관리의 용이성을 위해 모두 Int32로 변경   * 연관된 'Challenge/Project/Todo' Object 변수처리 변경

+++ teamplan/Sources/Object/Challenge/ChallengeObject.swift
@@ -13,7 +13,7 @@ import Foundation
 //============================
 struct ChallengeObject{
     // id
-    let chlg_id: Int64
+    let chlg_id: Int
     
     // category
     let chlg_type: ChallengeType
@@ -21,7 +21,7 @@ struct ChallengeObject{
     // content
     let chlg_title: String
     let chlg_desc: String
-    let chlg_goal: Int64
+    let chlg_goal: Int
     let chlg_reward: Int
     
     // status
@@ -38,11 +38,11 @@ struct ChallengeObject{
     // Constructor
     // Get Coredata
     init(chlgEntity: ChallengeEntity) {
-        self.chlg_id = chlgEntity.chlg_id
+        self.chlg_id = Int(chlgEntity.chlg_id)
         self.chlg_type = ChallengeType(rawValue: chlgEntity.chlg_type!) ?? .unkown
         self.chlg_title = chlgEntity.chlg_title!
         self.chlg_desc = chlgEntity.chlg_desc!
-        self.chlg_goal = chlgEntity.chlg_goal
+        self.chlg_goal = Int(chlgEntity.chlg_goal)
         self.chlg_reward = Int(chlgEntity.chlg_reward)
         self.chlg_step = Int(chlgEntity.chlg_step)
         self.chlg_selected = chlgEntity.chlg_selected
@@ -55,11 +55,11 @@ struct ChallengeObject{
     
     // Get Dummy
     init(chlg_id: Int64, chlg_type: ChallengeType, chlg_title: String, chlg_desc: String, chlg_goal: Int64, chlg_reward: Int, chlg_step: Int, chlg_selected: Bool, chlg_status: Bool, chlg_lock: Bool, chlg_selected_at: Date, chlg_unselected_at: Date, chlg_finished_at: Date) {
-        self.chlg_id = chlg_id
+        self.chlg_id = Int(chlg_id)
         self.chlg_type = chlg_type
         self.chlg_title = chlg_title
         self.chlg_desc = chlg_desc
-        self.chlg_goal = chlg_goal
+        self.chlg_goal = Int(chlg_goal)
         self.chlg_reward = chlg_reward
         self.chlg_step = chlg_step
         self.chlg_selected = chlg_selected
